Add initial lookup functionality to the bugjira library
Adds a new class, Bugjira, that acts as an abstraction layer for the
bugzilla.Bugzilla and jira.JIRA classes' functionality.

Allows lookup of bugzilla or jira issues from their respective remote services
using a single get_issue(key) method. Which service to query is determined by
the pattern of the key parameter, i.e. passing in a bug id will cause a
bugzilla query, and passing in a JIRA issue key will query the Jira service.

The returned bug (bz) or issue (jira) is wrapped in a new class: Issue. This
class will, in future versions, provide additional abstraction to allow
interacting with the respective remote object types.

In both cases, the remote objects that we are abstracting can be accessed
directly. Instances of Bugjira have attributes, bugzilla and jira, that are
pointers to the bugzilla.Bugzilla and jira.JIRA objects whose lookup methods
are what this patch attempts to abstract. Similarly, instances of Issue have
attributes, bugzilla and jira_issue, that are set conditionally to point to
either a bugzilla object or a jira issue (i.e. what we are actually getting
back from the backend api libraries). So if the abstraction provided by the
bugjira library is insufficient for your use case, you can switch directly to
using the bugzilla or jira libraries at will.