Rename ConfigDict class to avoid name clash with pydantic
The workaround in the previous patch left some technical debt in the form of
ambiguity in the class names. This patch corrects that by renaming our class
BugjiraConfigDict.

This patch also cleans up a pytest warning about how the pydantic
ValidationError was moved to a different namespace in that library by adjusting
the import statements we use.